national dashboard improvements
